requirements.txt
aiogram==3.17.0
aiohttp==3.11.11
python-dotenv==1.0.1
pillow==11.1.0